[mod-braille] Add step "css:extract"
... for extracting inline CSS, defined in attributes, into an external
stylesheet. Basically it's the inverse of css:inline.

I've chosen to implement the step in XProc but it could also have been
done in Java. For now the implementation is based on ID selectors. In
combination with css:inline, this step can be used to partly process a
style sheet. A more advanced solution could be thought of when styles
would be represented in an external XML document as described in
https://github.com/daisy/pipeline-mod-braille/issues/101.