[mod-braille] Change BrailleTranslatorResult contract
When force is false, nextTranslatedRow may return a string that is
shorter than countRemaining (possibly empty) even if countRemaining is
less than width. It can do this when it is desired to keep the last word
of a segment together with the next. Because a BrailleTranslatorResult
does not know anything about the surrounding segments (see issue
https://github.com/brailleapps/dotify.api/issues/13), it must assume
that the first word of the next segment is too wide. The
AbstractBrailleTranslator does this when it encounters ZWNBSP, which are
inserted by css-to-obfl.

The caller should take this new behavior into account otherwise it will
result in premature line breaks. After a nextTranslatedRow, the caller
should not break the line before having checked whether
getTranslatedRemainder of the current segment and nextTranslatedRow of
the next would fit on the remainder of the line.

Dotify has not been adapted yet.

see https://github.com/sbsdev/pipeline-mod-sbs/issues/63