Change the fake sections in the frontmatter to divs
The frontmatter typically contains level2 sections to separate the
front of the title page from the back. As these aren't real sections
they do not contain a header.

So the EPUB conversion adds fake headers which gives them more weight
that was ever intended.

To avoid this we turn these sections into a div before sending them to
the EPUB converter.

Fixes #14