Fix the ich tests
The reason I could not reproduce the issue report was due to the fact
that the tests cases weren't correct. The punctuation needs to be
outside the em

Port of sbsdev/dtbook2sbsform@dfc3e573f9144599e03d13515ff2e9f22697bd6d

The original commit apparently seems to have fixed a few things. But
as it stands this commit causes all the changed tests to fail. It also
seems a bit weird that only All-caps test inside emp should have to be
changed. Why are all other tests not updated? Why is the punctuation
inside the em for all other tests?