Add imgrefs automatically
Based on the assumption that we will markup every image in a separate
imggroup and that we will group images together with their respective
prodnote (long description) inside this imggroup, we can simply insert
imgrefs for those cases. This will simplify the processing in the
pipeline and we can rely on off the shelf functionality to create
EPUBs.